refactor: restructure UiTPAS cards components
